backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
groq==0.12.0
chromadb==0.5.23
sentence-transformers==3.3.1
reportlab==4.2.5
python-dotenv==1.0.1
sqlite-utils==3.38
sse-starlette==2.1.3
pydantic==2.10.4
pydantic-settings==2.6.1
python-multipart==0.0.12
httpx==0.27.2
pypdf==5.1.0
